Secbot with proven state sync property at level 1
This version of secbot contains the 'important' state invariant
Go => IDLE To prove this various other state invariants have to be added
and then extra guards are needed to say that we will not raise certain
internal triggers (send_message and spi_done) on transitions that would
violate these invariants. Some guards were also added manually (after
generation) to the future transition set events... again to say that we
will not raise internal triggers that violate the invariants in the
future. Eventually, these will be added automatically by the
translation.
Level 2 is not yet addressed in this commit